+++ src/pagination/pagination.jsx
@@ -133,6 +133,7 @@ class Pagination extends Component {
          * 弹层组件属性，透传给Popup
          */
         popupProps: PropTypes.object,
+        selectProps: PropTypes.object,
     };
 
     static defaultProps = {
@@ -476,7 +477,7 @@ class Pagination extends Component {
     }
 
     renderPageSizeDropdown() {
-        const { prefix, size, pageSizeList, locale, popupProps } = this.props;
+        const { prefix, size, pageSizeList, locale, popupProps, selectProps } = this.props;
         const { currentPageSize } = this.state;
 
         return (
@@ -485,8 +486,9 @@ class Pagination extends Component {
                 popupClassName={`${prefix}pagination-size-selector-popup`}
                 popupProps={popupProps}
                 aria-label={locale.selectAriaLabel}
-                size={size}
                 autoWidth={false}
+                selectProps={selectProps}
+                size={size}
                 value={currentPageSize}
                 onChange={this.onSelectSize.bind(this)}
             >
@@ -540,6 +542,7 @@ class Pagination extends Component {
             link,
             onChange,
             popupProps,
+            selectProps,
             ...others
         } = this.props;
         /* eslint-enable */