+++ src/step/view/step.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classNames from 'classnames';
 import React, { Component, Children } from 'react';
 import { polyfill } from 'react-lifecycles-compat';
-import { support, events, dom } from '../../util';
+import { support, events, dom, obj } from '../../util';
 
 const getHeight = el => dom.getStyle(el, 'height');
 const setHeight = (el, height) => dom.setStyle(el, 'height', height);
@@ -48,6 +48,10 @@ class Step extends Component {
          * @returns {Node} 节点的渲染结果
          */
         itemRender: PropTypes.func,
+        /**
+         * 宽度横向拉伸
+         */
+        stretch: PropTypes.bool,
     };
 
     static defaultProps = {
@@ -58,6 +62,7 @@ class Step extends Component {
         shape: 'circle',
         animation: true,
         itemRender: null,
+        stretch: false,
     };
 
     static contextTypes = {
@@ -183,17 +188,8 @@ class Step extends Component {
 
     render() {
         // eslint-disable-next-line
-        const {
-            className,
-            current,
-            labelPlacement,
-            shape,
-            readOnly,
-            animation,
-            itemRender,
-            rtl,
-            ...others
-        } = this.props;
+        const { className, current, labelPlacement, shape, readOnly, animation, itemRender, rtl, stretch } = this.props;
+        const others = obj.pickOthers(Step.propTypes, this.props);
         let { prefix, direction, children } = this.props;
         prefix = this.context.prefix || prefix;
         const { parentWidth, parentHeight } = this.state;
@@ -228,6 +224,7 @@ class Step extends Component {
                 onResize: () => {
                     this.step && this.adjustHeight();
                 },
+                stretch,
             });
         });
 
@@ -246,7 +243,6 @@ class Step extends Component {
         }
 
         // others.onKeyDown = makeChain(this.handleKeyDown, others.onKeyDown);
-
         return (
             <ol {...others} className={stepCls} ref={this._stepRefHandler}>
                 {cloneChildren}